ENGR00231256 EPDC: unit test: Incorrect display after screen mode change
The FBIOPUT_VSCREENINFO ioctl causes the fb console driver to clear
the FB to all zeros (black) when changing screen modes (RGB to grayscale).
Some tests did not account for this and so the display was incorrect
after executing the ioctl. Fix is to blank the screen to white and
redraw the contents (if necessary).

In addition, the following minor changes were made to the unit test:

1. Added max SEND_UPDATE retry so test doesn't hang if driver in bad state.
2. Don't check for collision status if wait for update fails.
3. test_auto_waveform: Don't check for waveform selected
   (not implemented by mx6 epdc driver).
4. test_animation_mode: blank screen on enter and exit from A2 animation mode.
5. test_fast_square: wait for updates so moving square can be seen.
6. test_dry_run: disallow running with snapshot scheme.
7. Add "-p" argument documentation and other minor changes in usage strings.
8. Add printfs when exiting due to error.

Signed-off-by: Michael Minnick <michael.minnick@freescale.com>