MLK-12208 v4l2 capture: Fix USB camera met "set dest crop failed" issue
The commit 4ee3c02f7669869fb49abbe2da955a92b3e5ea06 introduces another ioctl
for the crop feature for SoC camera, but does not consider the USB camera
case which does not support CROP feature. This patch added the condition
judgement to not apply it on USB camera to fix it.

Signed-off-by: Robby Cai <robby.cai@nxp.com>